add ajax

+++ DriverExamSystem/src/main/java/com/java4/des/entity/Student.java
@@ -19,7 +19,7 @@ public class Student {
 	private String stuName;			//学生姓名	
 	private String stuPass;			//密码
 	private int stuSex;			//性别
-	private float stuScore;			//成绩
+	private Integer stuScore;			//成绩
 	private String teaName;			//对应教练的姓名
 	
 }
